requirements.txt
asgiref==3.7.2
cfgv==3.4.0
distlib==0.3.8
Django==5.0.2
djangorestframework==3.14.0
filelock==3.13.1
identify==2.5.35
nodeenv==1.8.0
platformdirs==4.2.0
pre-commit==3.6.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
tzdata==2024.1
virtualenv==20.25.1
